# dotnet-video-center/ClientApp/src/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Media And Tag API",
    "description": "Manage your local media files.",
    "contact": {
      "name": "mushroom-cn",
      "url": "https://github.com/mushroom-cn/dotnet-video-center"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/mushroom-cn/dotnet-video-center/blob/main/LICENSE"
    },
    "version": "v1"
  },
  "paths": {
    "/v1/Media/{id}": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Get media description by media id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Media id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the media description.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          },
          "404": {
            "description": "Media description not found.",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Media"
        ],
        "summary": "Modify media description.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Media id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "Media entity",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Media"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Media"
        ],
        "summary": "Delete media description.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Media id.",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/v1/Media/ids": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Get media description collection by media ids.",
        "parameters": [
          {
            "name": "ids",
            "in": "query",
            "description": "A group of media id",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the media description.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Media description not found."
          }
        }
      },
      "put": {
        "tags": [
          "Media"
        ],
        "summary": "Batch modify media description.",
        "requestBody": {
          "description": "Media entity collection.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/Media/name/{name}": {
      "get": {
        "tags": [
          "Media"
        ],
        "summary": "Get media description collection by media names. support flex search",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "description": "Media name",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returns the media description.",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Media description not found."
          }
        }
      }
    },
    "/v1/Media": {
      "post": {
        "tags": [
          "Media"
        ],
        "summary": "Batch create media description.",
        "requestBody": {
          "description": "Media entity collection.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Media"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Media"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Media"
        ],
        "summary": "Batch delete media description.",
        "requestBody": {
          "description": "A group of media id.",
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "text/json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "application/*+json": {
              "schema": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/v1/Media/convert/{id}": {
      "put": {
        "tags": [
          "Media"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Media": {
        "required": [
          "iconUrl",
          "mediaType",
          "name",
          "path"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 250,
            "type": "string"
          },
          "path": {
            "maxLength": 300,
            "type": "string"
          },
          "mediaType": {
            "maxLength": 30,
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifyTime": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "maxLength": 200,
            "type": "string",
            "nullable": true
          },
          "mediaTags": {
            "maxItems": 10,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaTag"
            },
            "nullable": true
          },
          "iconUrl": {
            "type": "string"
          },
          "url": {
            "maxLength": 2083,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "MediaTag": {
        "required": [
          "mediaId",
          "tagId"
        ],
        "type": "object",
        "properties": {
          "publicationDate": {
            "type": "string",
            "format": "date-time"
          },
          "tagId": {
            "type": "string",
            "format": "uuid"
          },
          "tag": {
            "$ref": "#/components/schemas/Tag"
          },
          "mediaId": {
            "type": "string",
            "format": "uuid"
          },
          "media": {
            "$ref": "#/components/schemas/Media"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Tag": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "maxLength": 250,
            "type": "string"
          },
          "createTime": {
            "type": "string",
            "format": "date-time"
          },
          "lastModifyTime": {
            "type": "string",
            "format": "date-time"
          },
          "mediaTags": {
            "maxItems": 999,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MediaTag"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      }
    }
  }
}